Updated Settings
- split settings::LocalComponent into Audio, Camera and Window
- updated audio settings to be stored as 0 to 1 instead of 0 to 100
- hooked up master volume to music and sfx

+++ Code/Application/GameUI/GameUI/SettingsMenuComponents.h
@@ -14,10 +14,10 @@ namespace gui::settings_menu
 	struct ValueRequest final : public ecs::EventComponent<ValueRequest> 
 	{ 
 		// audio
-		Optional<int32> m_EffectVolume = {};
-		Optional<int32> m_MasterVolume = {};
-		Optional<int32> m_MusicVolume = {};
-		// gameplay
+		Optional<float> m_EffectVolume = {};
+		Optional<float> m_MasterVolume = {};
+		Optional<float> m_MusicVolume = {};
+		// camera
 		Optional<float> m_MoveSpeed = {};
 		Optional<float> m_ZoomRate = {};
 		Optional<float> m_ZoomSpeed = {};
